requirements.txt
flask==3.1.0
torch==2.6.0
transformers==4.48.2
faiss-cpu==1.10.0
pandas==2.2.3